rename plural directories / files
