+++ internal/services/phone/callqueuemember/call_queue_member_crud.go
@@ -1,4 +1,4 @@
-package callqueuemembers
+package callqueuemember
 
 import (
 	"context"
