+++ internal/services/phone/callqueuemember/call_queue_member_dto.go
@@ -1,4 +1,4 @@
-package callqueuemembers
+package callqueuemember
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
