+++ internal/services/phone/callqueuemember/call_queue_members_resource.go
@@ -1,4 +1,4 @@
-package callqueuemembers
+package callqueuemember
 
 import (
 	"context"
