+++ internal/services/phone/callqueuephonenumber/call_queue_phone_number_crud.go
@@ -1,4 +1,4 @@
-package callqueuephonenumbers
+package callqueuephonenumber
 
 import (
 	"context"
