+++ internal/services/phone/callqueuephonenumber/call_queue_phone_numbers_resource.go
@@ -1,4 +1,4 @@
-package userphonenumbers
+package callqueuephonenumber
 
 import (
 	"context"
@@ -6,9 +6,13 @@ import (
 	"strings"
 
 	"github.com/folio-sec/terraform-provider-zoom/internal/provider/shared"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/samber/lo"
@@ -20,7 +24,7 @@ var (
 	_ resource.ResourceWithImportState = &tfResource{}
 )
 
-func NewPhoneUserPhoneNumbersResource() resource.Resource {
+func NewPhoneCallQueuePhoneNumbersResource() resource.Resource {
 	return &tfResource{}
 }
 
@@ -44,23 +48,27 @@ func (r *tfResource) Configure(_ context.Context, req resource.ConfigureRequest,
 }
 
 func (r *tfResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_phone_user_phone_numbers"
+	resp.TypeName = req.ProviderTypeName + "_phone_call_queue_phone_numbers"
 }
 
 func (r *tfResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		MarkdownDescription: `Assigns a [phone number](https://support.zoom.us/hc/en-us/articles/360020808292-Managing-Phone-Numbers) to a user who has already enabled Zoom Phone.
+		MarkdownDescription: `After [buying phone number(s)](https://support.zoom.us/hc/en-us/articles/360020808292#h_007ec8c2-0914-4265-8351-96ab23efa3ad), you can assign it, allowing callers to directly dial a number to reach a [call queue](https://support.zoom.us/hc/en-us/articles/360021524831-Managing-Call-Queues).
 
 ## API Permissions
 The following API permissions are required in order to use this resource.
 This resource requires the ` + strings.Join([]string{
-			"`phone:write:user_number:admin`",
-			"`phone:delete:user_number:admin`",
+			"`phone:read:call_queue:admin`",
+			"`phone:read:list_call_queues:admin`",
+			"`phone:read:list_numbers:admin`",
+			"`phone:write:call_queue_number:admin`",
+			"`phone:delete:call_queue_number:admin`",
 		}, ", ") + ".",
 		Attributes: map[string]schema.Attribute{
-			"user_id": schema.StringAttribute{
+			"call_queue_id": schema.StringAttribute{
 				Required:            true,
-				MarkdownDescription: "Unique identifier of the User.",
+				MarkdownDescription: "Unique identifier of the Call Queue.",
+				PlanModifiers:       []planmodifier.String{stringplanmodifier.RequiresReplace()},
 			},
 			"phone_numbers": schema.SetNestedAttribute{
 				Required: true,
@@ -74,7 +82,16 @@ This resource requires the ` + strings.Join([]string{
 						"number": schema.StringAttribute{
 							Optional:            true,
 							Computed:            true,
-							MarkdownDescription: "Phone number e.g. `+12058945456` . Provide either the `id` or the `number` field. ",
+							MarkdownDescription: "Phone number e.g. `+12058945456`. Provide either the `id` or the `number` field. ",
+						},
+						"source": schema.StringAttribute{
+							Optional: true,
+							Computed: true,
+							MarkdownDescription: `Source
+  - Allowed: internal┃external`,
+							Validators: []validator.String{
+								stringvalidator.OneOf("internal", "external"),
+							},
 						},
 					},
 				},
@@ -84,13 +101,14 @@ This resource requires the ` + strings.Join([]string{
 }
 
 type resourceModel struct {
-	UserID       types.String                `tfsdk:"user_id"`
+	CallQueueID  types.String                `tfsdk:"call_queue_id"`
 	PhoneNumbers []*resourceModelPhoneNumber `tfsdk:"phone_numbers"`
 }
 
 type resourceModelPhoneNumber struct {
 	ID     types.String `tfsdk:"id"`
 	Number types.String `tfsdk:"number"`
+	Source types.String `tfsdk:"source"`
 }
 
 func (r *tfResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
@@ -103,7 +121,7 @@ func (r *tfResource) Read(ctx context.Context, req resource.ReadRequest, resp *r
 
 	output, err := r.read(ctx, state)
 	if err != nil {
-		resp.Diagnostics.AddError("Error reading phone user phone numbers", err.Error())
+		resp.Diagnostics.AddError("Error reading phone call queue phone numbers", err.Error())
 		return
 	}
 
@@ -115,7 +133,7 @@ func (r *tfResource) Read(ctx context.Context, req resource.ReadRequest, resp *r
 }
 
 func (r *tfResource) read(ctx context.Context, plan resourceModel) (*resourceModel, error) {
-	dto, err := r.crud.read(ctx, plan.UserID)
+	dto, err := r.crud.read(ctx, plan.CallQueueID)
 	if err != nil {
 		return nil, err
 	}
@@ -127,10 +145,11 @@ func (r *tfResource) read(ctx context.Context, plan resourceModel) (*resourceMod
 		return &resourceModelPhoneNumber{
 			ID:     p.id,
 			Number: p.number,
+			Source: p.source,
 		}
 	})
 	return &resourceModel{
-		UserID:       plan.UserID,
+		CallQueueID:  plan.CallQueueID,
 		PhoneNumbers: phoneNumbers,
 	}, nil
 }
@@ -141,7 +160,7 @@ func (r *tfResource) sync(ctx context.Context, plan resourceModel) error {
 		return err
 	}
 	if asis == nil {
-		return fmt.Errorf("phone user not found %s", plan.UserID.ValueString())
+		return fmt.Errorf("calal queue not found %s", plan.CallQueueID.ValueString())
 	}
 
 	// 0. plan validation (it might be better to move into validator)
@@ -163,7 +182,7 @@ func (r *tfResource) sync(ctx context.Context, plan resourceModel) error {
 		}
 	}
 	if err = r.crud.unassign(ctx, &unassignDto{
-		userID:         plan.UserID,
+		callQueueID:    plan.CallQueueID,
 		phoneNumberIDs: unassignPhoneNumberIDs,
 	}); err != nil {
 		return err
@@ -186,7 +205,7 @@ func (r *tfResource) sync(ctx context.Context, plan resourceModel) error {
 		}
 	}
 	if err = r.crud.assign(ctx, &assignDto{
-		userID:         plan.UserID,
+		callQueueID:    plan.CallQueueID,
 		phoneNumberIDs: assignPhoneNumberIDs,
 		phoneNumbers:   assignPhoneNumbers,
 	}); err != nil {
@@ -205,7 +224,7 @@ func (r *tfResource) Create(ctx context.Context, req resource.CreateRequest, res
 
 	if err := r.sync(ctx, plan); err != nil {
 		resp.Diagnostics.AddError(
-			"Error creating phone user phone numbers",
+			"Error creating phone call queue phone numbers",
 			err.Error(),
 		)
 		return
@@ -213,7 +232,7 @@ func (r *tfResource) Create(ctx context.Context, req resource.CreateRequest, res
 
 	output, err := r.read(ctx, plan)
 	if err != nil {
-		resp.Diagnostics.AddError("Error creating phone user phone numbers on reading", err.Error())
+		resp.Diagnostics.AddError("Error creating phone call queue phone numbers on reading", err.Error())
 		return
 	}
 
@@ -230,15 +249,15 @@ func (r *tfResource) Update(ctx context.Context, req resource.UpdateRequest, res
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		resp.Diagnostics.AddError(
-			"Error updating phone user phone numbers on get plan",
-			"Error updating phone user phone numbers",
+			"Error updating phone call queue phone numbers on get plan",
+			"Error updating phone call queue phone numbers",
 		)
 		return
 	}
 
 	if err := r.sync(ctx, plan); err != nil {
 		resp.Diagnostics.AddError(
-			"Error creating phone user phone numbers",
+			"Error creating phone call queue phone numbers",
 			err.Error(),
 		)
 		return
@@ -246,7 +265,7 @@ func (r *tfResource) Update(ctx context.Context, req resource.UpdateRequest, res
 
 	output, err := r.read(ctx, plan)
 	if err != nil {
-		resp.Diagnostics.AddError("Error updating phone user phone numbers", err.Error())
+		resp.Diagnostics.AddError("Error updating phone call queue phone numbers", err.Error())
 		return
 	}
 
@@ -265,45 +284,23 @@ func (r *tfResource) Delete(ctx context.Context, req resource.DeleteRequest, res
 		return
 	}
 
-	asis, err := r.read(ctx, state)
-	if err != nil {
+	if err := r.crud.unassignAll(ctx, state.CallQueueID); err != nil {
 		resp.Diagnostics.AddError(
-			"Error deleting phone user phone numbers on read",
+			"Error deleting phone call queue phone numbers",
 			fmt.Sprintf(
-				"Could not delete phone user phone numbers %s, unexpected error: %s",
-				state.UserID.ValueString(),
-				err,
-			),
-		)
-		return
-	}
-	if asis == nil {
-		return // already deleted
-	}
-
-	dto := &unassignDto{
-		userID: state.UserID,
-		phoneNumberIDs: lo.Map(asis.PhoneNumbers, func(p *resourceModelPhoneNumber, _index int) types.String {
-			return p.ID
-		}),
-	}
-	if err := r.crud.unassign(ctx, dto); err != nil {
-		resp.Diagnostics.AddError(
-			"Error deleting phone user phone numbers",
-			fmt.Sprintf(
-				"Could not delete phone user phone numbers %s, unexpected error: %s",
-				state.UserID.ValueString(),
+				"Could not delete phone call queue phone numbers %s, unexpected error: %s",
+				state.CallQueueID.ValueString(),
 				err,
 			),
 		)
 		return
 	}
 
-	tflog.Info(ctx, "deleted phone user phone numbers", map[string]interface{}{
-		"user_id": state.UserID.ValueString(),
+	tflog.Info(ctx, "deleted phone call queue phone numbers", map[string]interface{}{
+		"call_queue_id": state.CallQueueID.ValueString(),
 	})
 }
 
 func (r *tfResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
-	resource.ImportStatePassthroughID(ctx, path.Root("user_id"), req, resp)
+	resource.ImportStatePassthroughID(ctx, path.Root("call_queue_id"), req, resp)
 }
