+++ internal/services/phone/userphonenumber/user_phone_number_crud.go
@@ -1,4 +1,4 @@
-package userphonenumbers
+package userphonenumber
 
 import (
 	"context"
