+++ internal/services/phone/userphonenumber/user_phone_number_dto.go
@@ -1,4 +1,4 @@
-package userphonenumbers
+package userphonenumber
 
 import "github.com/hashicorp/terraform-plugin-framework/types"
